requirements.txt
Flask
mysql.connector
flask_cors